Address review comments: deduplicate helpers, add validation, fix import
- Extract getResourceTypeSegment/getLastPathSegment to utils.ts (comment #1)
- Add validation for singleton/variable segments in getResourceTypeSegment (comment #2)
- Add missing NonResourceMethod type import in test file (comment #3)
